% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example3_alpha_next_function.R
\name{Example3_alpha_next_function}
\alias{Example3_alpha_next_function}
\title{Title}
\usage{
Example3_alpha_next_function(alpha, P, Y_T, N, K)
}
\arguments{
\item{alpha}{alpha vector (intercept, coefficients and error variance) from the last iteration}

\item{P}{Transpose of the transition matrix Pi}

\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
Computes,for Example 3 setups (All Switching), the alpha matrix for the next iteration, based on the last iteration.
}
\description{
Title
}
\examples{
Example3_alpha_next_function(alpha,P,Y_T,N,K)
}
