% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example3_EM_Algorithm_function_lastbeta.R
\name{Example3_EM_Algorithm_function_lastbeta}
\alias{Example3_EM_Algorithm_function_lastbeta}
\title{Title}
\usage{
Example3_EM_Algorithm_function_lastbeta(
  Y_T,
  K,
  N,
  m,
  threshold = 0.5,
  max = 500,
  R = 15,
  Crit = "LV"
)
}
\arguments{
\item{Y_T}{Data}

\item{K}{Lag order of the AR-Model}

\item{N}{Number of underlying Regimes}

\item{m}{Number of observations to condition on}

\item{threshold}{Threshold for assigning the underlying regime based on the estimated probability}

\item{max}{Number of iterations to go through}

\item{R}{Number of random starting points}

\item{Crit}{Performance Metric to use for choosing the optimisation result}
}
\value{
This function brings, for Example 3 setups (All Switching), everything together and runs the EM-Algorithm
}
\description{
Title
}
\examples{
Example3_EM_Algorithm_function_lastbeta(Y_T,K,N,m,threshold = 0.5,max = 500,R = 15, Crit = "LV")
}
