% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example4_alpha_next_function_lastbeta.R
\name{Example4_alpha_next_function_lastbeta}
\alias{Example4_alpha_next_function_lastbeta}
\title{Example4_alpha_next_function_lastbeta
Example4_alpha_next_function_lastbeta allows its user to estimate the alpha matrix of the next iteration for Example 4 setups, based on Müller (2025, page 19-21)}
\usage{
Example4_alpha_next_function_lastbeta(Y_T, alpha, P, Switcher, N, K, m)
}
\arguments{
\item{Y_T}{Data}

\item{alpha}{alpha matrix from the last iteration}

\item{P}{P matrix from the last iteration}

\item{Switcher}{Switching vector which indicates which parameter switch}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional log-likelihood conditions on}
}
\value{
Returns a matrix of the dimensions (N x K+2)
}
\description{
Example4_alpha_next_function_lastbeta
Example4_alpha_next_function_lastbeta allows its user to estimate the alpha matrix of the next iteration for Example 4 setups, based on Müller (2025, page 19-21)
}
\examples{
Example4_alpha_next_function_lastbeta(Y_T,alpha,P,Switcher,N,K,m)
}
