% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example4_EM_Algorithm_function_lastbeta.R
\name{Example4_EM_Algorithm_function_lastbeta}
\alias{Example4_EM_Algorithm_function_lastbeta}
\title{Example4_EM_Algorithm_function_lastbeta
Example4_EM_Algorithm_function_lastbeta allows its user to run the EM algorithm for estimating the parameters of a Markov-Switching AR model of the form of Example 4, see Müller (2025, page 19-21)}
\usage{
Example4_EM_Algorithm_function_lastbeta(
  Y_T,
  K,
  N,
  m,
  threshold = 0.5,
  max = 500,
  R = 15,
  Crit = "LV",
  all.plot = FALSE,
  Switcher
)
}
\arguments{
\item{Y_T}{Data}

\item{K}{Lag-order of the AR model}

\item{N}{Number of underlying regimes}

\item{m}{Number of time periods the conditional log-likelihood conditions on}

\item{threshold}{Threshold for assigning the regimes}

\item{max}{Number of iterations of the EM algorithm}

\item{R}{Number of random starting points of the EM algorithm}

\item{Crit}{Performance metric used for choosing between the optima}

\item{all.plot}{Whether the plots from all optimisation attempts should be printed}

\item{Switcher}{Switching vector indicating for all parameters whether they switch or not}
}
\value{
Returns a list object containing the results from the EM algorithm estimation of the Markov-Switching AR model
}
\description{
Example4_EM_Algorithm_function_lastbeta
Example4_EM_Algorithm_function_lastbeta allows its user to run the EM algorithm for estimating the parameters of a Markov-Switching AR model of the form of Example 4, see Müller (2025, page 19-21)
}
\examples{
Example4_EM_Algorithm_function_lastbeta(Y_T,K,N,m,threshold = 0.5,max = 500,R = 15, Crit = "LV", all.plot = FALSE, Switcher)
}
