% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example5_alpha_next_function_lastbeta.R
\name{Example5_alpha_next_function_lastbeta}
\alias{Example5_alpha_next_function_lastbeta}
\title{Example5_alpha_next_function_lastbeta
Example5_alpha_next_function_lastbeta allows its user to estimate the alpha matrix of the next iteration as part of the EM algorithm for setups of the form of Example 5. A more detailed explanation can be found in Müller (2025, page 21-24)}
\usage{
Example5_alpha_next_function_lastbeta(Y_T, alpha, P, Switcher, N, K, m)
}
\arguments{
\item{Y_T}{Data}

\item{alpha}{alpha matrix from the previous iteration}

\item{P}{P matrix from the previous iteration}

\item{Switcher}{Switching vector indicating which parameters switch}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional likelihood conditions on}
}
\value{
Returns a matrix of the dimensions (N x K+2)
}
\description{
Example5_alpha_next_function_lastbeta
Example5_alpha_next_function_lastbeta allows its user to estimate the alpha matrix of the next iteration as part of the EM algorithm for setups of the form of Example 5. A more detailed explanation can be found in Müller (2025, page 21-24)
}
\examples{
Example5_alpha_next_function_lastbeta(Y_T,alpha,P,Switcher,N,K,m)
}
