% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example5_Beta_F_function.R
\name{Example5_Beta_F_function}
\alias{Example5_Beta_F_function}
\title{Example5_Beta_F_function
Example5_Beta_F_function allows its user to compute beta_F as part of the EM algorithm for Example 5 setups, for details see Müller (2025, page 23-24)}
\usage{
Example5_Beta_F_function(Y_T, alpha, P, Switcher, N, K, m)
}
\arguments{
\item{Y_T}{Data}

\item{alpha}{alpha matrix from the previous iteration}

\item{P}{P matrix from the previous iteration}

\item{Switcher}{Switching vector indicating whether a parameter is fixed or switches}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional likelihood conditions on}
}
\value{
Returns a vector representing the estimates for beta_F based on the results from the last iteration
}
\description{
Example5_Beta_F_function
Example5_Beta_F_function allows its user to compute beta_F as part of the EM algorithm for Example 5 setups, for details see Müller (2025, page 23-24)
}
\examples{
Example5_Beta_F_function(Y_T,alpha,P,Switcher,N,K,m)
}
