% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example5_EM_Algorithm_function_lastbeta.R
\name{Example5_EM_Algorithm_function_lastbeta}
\alias{Example5_EM_Algorithm_function_lastbeta}
\title{Example5_EM_Algorithm_function_lastbeta
Example5_EM_Algorithm_function_lastbeta allows its user to implement the EM algorithm for estimating the parameters in Example 5 setups, see Müller (2025, page 21-24)}
\usage{
Example5_EM_Algorithm_function_lastbeta(
  Y_T,
  K,
  N,
  m,
  threshold = 0.5,
  max = 500,
  R = 15,
  Crit = "LV",
  all.plot = FALSE,
  Switcher
)
}
\arguments{
\item{Y_T}{Data}

\item{K}{Lag-order of the AR model}

\item{N}{Number of underlying regimes}

\item{m}{Number of time periods the conditional likelihood conditions on}

\item{threshold}{Threshold for assigning the predicted regimes}

\item{max}{Number of iterations of the EM algorithm}

\item{R}{Number of random starting points for the EM algorithm}

\item{Crit}{Performance metric for selecting one of the optima}

\item{all.plot}{Whether plots from a optimisation attempts should be printed}

\item{Switcher}{Switching vector indicating for all parameters whether they switch or not}
}
\description{
Example5_EM_Algorithm_function_lastbeta
Example5_EM_Algorithm_function_lastbeta allows its user to implement the EM algorithm for estimating the parameters in Example 5 setups, see Müller (2025, page 21-24)
}
\examples{
Example5_EM_Algorithm_function_lastbeta(Y_T,K,N,m,threshold = 0.5,max = 500,R = 15, Crit = "LV", all.plot = FALSE, Switcher)
}
