% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example5_sigma_function.R
\name{Example5_sigma_function}
\alias{Example5_sigma_function}
\title{Example5_sigma_function
Example5_sigma_function allows its user to esimate the error term standard deviation for Example 5 setups, based on the results from previous iterations and the formula presented in Müller (2025, page 24)}
\usage{
Example5_sigma_function(Y_T, alpha, P, Switcher, N, K, m)
}
\arguments{
\item{Y_T}{Data}

\item{alpha}{alpha matrix from the previous iteration}

\item{P}{P matrix from the previous iteration}

\item{Switcher}{Switching vector indicating whether a parameter switches or not}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional likelihood conditions on}
}
\description{
Example5_sigma_function
Example5_sigma_function allows its user to esimate the error term standard deviation for Example 5 setups, based on the results from previous iterations and the formula presented in Müller (2025, page 24)
}
\examples{
Example5_sigma_function(Y_T,alpha,P,Switcher,N,K,m)
}
