% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Log_Likelihood_function.R
\name{Log_Likelihood_function}
\alias{Log_Likelihood_function}
\title{Title}
\usage{
Log_Likelihood_function(alpha, P, Y_T, N, K)
}
\arguments{
\item{alpha}{alpha vector (intercept, coefficients and error variance) from the last iteration}

\item{P}{Transpose of the transition matrix Pi}

\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
Evaluates the conditional log-likelihood function
}
\description{
Title
}
\examples{
Log_Likelihood_function(alpha,P,Y_T,N,K)
}
