% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSARM_fit.R
\name{MSARM.fit}
\alias{MSARM.fit}
\title{MSARM.fit
MSARM.fit allows its user to estimate Markov-Switching AR models, as they are described in Hamilton (1994, page 690-698). Thereby an arbitrary set of parameters
is allowed to switch. The estimation process strongly orients itself on the EM algorithm described in Hamilton (1990) and is described in more detail in Müller (2025, page 7-21).}
\usage{
MSARM.fit(
  Y_T,
  K,
  N,
  Switcher,
  m = 1,
  threshold_value = 0.5,
  max_value = 250,
  R_value = 5,
  Crit_value = "LV",
  all.plot = FALSE
)
}
\arguments{
\item{Y_T}{Data (The full time series that is to be analyzed)}

\item{K}{Lag-order of the AR model}

\item{N}{Number of underlying Regimes}

\item{Switcher}{A vector of TRUE/FALSE indicating which parameters switch (intercept, coefficients, error term variance), the vector should have the length K+2}

\item{m}{Number of time periods the conditional likelihood function conditions on (generally set to 1)}

\item{threshold_value}{Threshold for assigning the estimated regimes (generally set to 0.5)}

\item{max_value}{Number of iterations for the EM algorithm (a higher iteration number increases the estimation quality)}

\item{R_value}{Number of random starting points for the EM algorithm (a higher number of random starting points increases estimation quality)}

\item{Crit_value}{Performance metric for choosing between the different optima. It is advised to run the function ones with "LV" and once with "RCM", if both estimation approaches lead to vastly different results a third run with Entropy can help to decide.}

\item{all.plot}{Whether plots from all optimisation attempts should be printed}
}
\value{
Returns a list object containing the estimation results regarding the Markov-Switching AR model. The output includes
the estimate of the P matrix, the estimate of the alpha matrix, the smoothed inference regarding the regines, the value of the log-likelihood function
the computed RSS value, the computed RCM value and the computed Entropy value. Furthermore the list contains the original time series, the in-sample fit and the residuals.
}
\description{
MSARM.fit
MSARM.fit allows its user to estimate Markov-Switching AR models, as they are described in Hamilton (1994, page 690-698). Thereby an arbitrary set of parameters
is allowed to switch. The estimation process strongly orients itself on the EM algorithm described in Hamilton (1990) and is described in more detail in Müller (2025, page 7-21).
}
\examples{
MSARM.fit(Y_T,K,N,Switcher) would be the standard implementation of MSARM.fit
}
