% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSARM_plot.R
\name{MSARM.plot}
\alias{MSARM.plot}
\title{MSARM.plot}
\usage{
MSARM.plot(res_MSARM.predict, conf = FALSE, start = c(1, 1), freq = 1)
}
\arguments{
\item{res_MSARM.predict}{An object saving the results from MSARM.predict}

\item{conf}{Set to TRUE if MSARM.predict was implemented with the option boot = TRUE, then bootstrap confidence intervals are plotted}

\item{start}{Start of the time series (format as in in the window() command), therefore (year,quarter) for quarterly data for example}

\item{freq}{Frequency of the data, i.e. 4 for quarterly, 12 for monthly data.}
}
\value{
Returns forecast plots
}
\description{
MSARM.plot
}
\examples{
MSARM.plot(res_MSARM.predict, start = c(1967,1), freq = 4) would give a standard forecast plot with the original data starting in 1967, first quarter
}
