% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSARM_predict.R
\name{MSARM.predict}
\alias{MSARM.predict}
\title{MSARM.predict}
\usage{
MSARM.predict(
  erg_MSARM.fit,
  n.ahead,
  boot = FALSE,
  levels = c(0.95, 0.9, 0.8, 0.7, 0.6),
  L = 10000
)
}
\arguments{
\item{erg_MSARM.fit}{Result of an MSARM.fit estimation}

\item{n.ahead}{Number of periods to be forecasted ahead}

\item{boot}{Whether bootstrapping shall be utilized (allows for creating confidence intervals)}

\item{levels}{If bootstrapped, the levels for the confidence intervals}

\item{L}{If bootstrapped the number of bootstrap estimates}
}
\value{
MSARM.predict allows its user to predict an MSARM.fit output for n.ahead periods ahead. Furthermore confidence intervals can be created by setting
boot = TRUE, in that case L bootstrap estimates and intervals will be created.
}
\description{
MSARM.predict
}
\examples{
MSARM.predict(erg_MSARM.fit, n.ahead, boot = FALSE, levels = c(0.95,0.9,0.8,0.7,0.6), L = 10000)
}
