% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_0_function.R
\name{P_0_function}
\alias{P_0_function}
\title{P_0_function}
\usage{
P_0_function(Y_T, N, K)
}
\arguments{
\item{Y_T}{Data}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}
}
\value{
Returns a matrix of the dimensions (N x N)
}
\description{
P_0_function creates P_0 as part of a random starting point for the EM algorithm
}
\examples{
P_0_function(Y_T,N,K)
}
