% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_ij_next_function.R
\name{P_ij_next_function}
\alias{P_ij_next_function}
\title{P_ij_next_function
P_ij_next_function allows its user to estimate the P matrix of the next iteration as part of the EM algorithm, as described in Hamilton (1994, page 695)}
\usage{
P_ij_next_function(alpha, P, Y_T, N, K, m)
}
\arguments{
\item{alpha}{alpha matrix of the last iteration}

\item{P}{P matrix of the last iteration}

\item{Y_T}{Data}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional likelihood function conditions on}
}
\value{
Returns a matrix of the dimensions (N x N)
}
\description{
P_ij_next_function
P_ij_next_function allows its user to estimate the P matrix of the next iteration as part of the EM algorithm, as described in Hamilton (1994, page 695)
}
\examples{
P_ij_next_function(alpha,P,Y_T,N,K,m)
}
