% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_Yt_function.R
\name{zeta_Yt_function}
\alias{zeta_Yt_function}
\title{Title}
\usage{
zeta_Yt_function(alpha, P, Y_T, N, K)
}
\arguments{
\item{alpha}{alpha vector (intercept, coefficients and error variance) from the last iteration}

\item{P}{Transpose of the transition matrix Pi}

\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
Estimates the optimal inference (Bayes Formula) about regimes
}
\description{
Title
}
\examples{
zeta_Yt_function(alpha,P,Y_T,N,K)
}
