% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_0_function.R
\name{alpha_0_function}
\alias{alpha_0_function}
\title{alpha_0_function}
\usage{
alpha_0_function(Y_T, N, K)
}
\arguments{
\item{Y_T}{Data}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}
}
\value{
Returns a matirx of the dimensions (N x K+2) containing randomly generated values for (intercept, coefficients, error term variance)
}
\description{
A random starting value of alpha is generated so that the EM algorithm can start
}
\examples{
alpha_0_function(Y_T,N,K)

}
