% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_0_function.R
\name{alpha_0_function}
\alias{alpha_0_function}
\title{alpha_0_function}
\usage{
alpha_0_function(Y_T, N, K)
}
\arguments{
\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
A random starting value of alpha is generated
}
\description{
alpha_0_function
}
\examples{
alpha_0_function(Y_T,N,K)
}
