% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example3_alpha_next_function.R
\name{Example3_alpha_next_function}
\alias{Example3_alpha_next_function}
\title{Example3_alpha_next_function}
\usage{
Example3_alpha_next_function(alpha, P, Y_T, N, K, m)
}
\arguments{
\item{alpha}{alpha vector (intercept, coefficients and error variance) from the last iteration}

\item{P}{transpose of the transition matrix (of the Markov-Chain) from the last iteration}

\item{Y_T}{Data}

\item{N}{Number of regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional likelihood conditions on}
}
\value{
returns a matrix of the dimensions (N x K+2)
}
\description{
The next alpha for Example 3 setups (all parameters switch) is computed as part of the EM algorithm
}
\examples{
Example3_alpha_next_function(alpha,P,Y_T,N,K,m)
}
