% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example3_EM_Algorithm_function_lastbeta.R
\name{Example3_EM_Algorithm_function_lastbeta}
\alias{Example3_EM_Algorithm_function_lastbeta}
\title{Example3_EM_Algorithm_function_lastbeta
Example3_EM_Algorithm_function_lastbeta allows its user to apply the EM algorithm for estimating Markov-Switching models of the form of Example 3 (All Switching), as described in Müller (2025, page 18-19)}
\usage{
Example3_EM_Algorithm_function_lastbeta(
  Y_T,
  K,
  N,
  m,
  threshold = 0.5,
  max = 500,
  R = 15,
  Crit = "LV",
  all.plot = FALSE,
  Switcher
)
}
\arguments{
\item{Y_T}{Data}

\item{K}{Lag-order of the AR model}

\item{N}{Number of underlying regimes}

\item{m}{Number of time periods the conditional likelihood conditions on}

\item{threshold}{#threshold for assigning the predicted regimes}

\item{max}{#number of iterations for the EM algorithm}

\item{R}{#number of random starting points for the EM algorithm}

\item{Crit}{#performance metric for choosing one of the R optima}

\item{all.plot}{#whether plots from all R optimisation attempts should be printed}

\item{Switcher}{#Switching vector indicating which parameters switch}
}
\description{
Example3_EM_Algorithm_function_lastbeta
Example3_EM_Algorithm_function_lastbeta allows its user to apply the EM algorithm for estimating Markov-Switching models of the form of Example 3 (All Switching), as described in Müller (2025, page 18-19)
}
\examples{
Example3_EM_Algorithm_function_lastbeta(Y_T,K,N,m,threshold = 0.5,max = 500,R = 15, Crit = "LV", all.plot = FALSE, Switcher)
}
