% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example4_Beta_F_function.R
\name{Example4_Beta_F_function}
\alias{Example4_Beta_F_function}
\title{Example4_Beta_F_function}
\usage{
Example4_Beta_F_function(Y_T, alpha, P, Switcher = Switcher, N, K, m)
}
\arguments{
\item{Y_T}{Data}

\item{alpha}{alpha matrix from the previous iteration}

\item{P}{P matrix from the previous iteration}

\item{Switcher}{Switching vector indicating which parameters switch}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}

\item{m}{Number of time periods the conditional log-likelihood conditions on}
}
\value{
Returns a vector representing beta_F based on the results from the last iteration
}
\description{
Example4_Beta_F_function allows its user to estimate the fixed coefficients vector called beta_F, based on the results from the previous iteration and the formulas in Müller (2025, page 21)
}
\examples{
Example4_Beta_F_function(Y_T,alpha,P,Switcher = Switcher,N,K,m)
}
