% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example5_Beta_F_function.R
\name{Example5_Beta_F_function}
\alias{Example5_Beta_F_function}
\title{Title}
\usage{
Example5_Beta_F_function(Y_T, alpha, P, Switcher, N, K, m)
}
\arguments{
\item{Y_T}{Data}

\item{alpha}{alpha vector (intercept, coefficients and error variance) from the last iteration}

\item{P}{Transpose of the transition matrix Pi}

\item{Switcher}{Switching-vector, a vector of (TRUE/FALSE) of the length K+2, which indicates which parameters are supposed to switch}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\description{
Title
}
\examples{
Example5_Beta_F_function(Y_T,alpha,P,Switcher,N,K)
}
