% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSARM_fit.R
\name{MSARM.fit}
\alias{MSARM.fit}
\title{MSARM.fit}
\usage{
MSARM.fit(
  Y_T,
  K,
  N,
  Switcher,
  m = 1,
  threshold_value = 0.5,
  max_value = 250,
  R_value = 5,
  Crit_value = "LV"
)
}
\arguments{
\item{Y_T}{Data}

\item{K}{Lag order of the AR-Model}

\item{N}{Number of underlying Regimes}

\item{Switcher}{Switching-vector, a vector of (TRUE/FALSE) of the length K+2, which indicates which parameters are supposed to switch}

\item{m}{Number of observations to condition on (generally set to 1)}

\item{threshold_value}{Threshold for assigning the underlying regime based on the estimated probability}

\item{max_value}{Number of iterations to go through}

\item{R_value}{Number of random starting points for the optimisation}

\item{Crit_value}{Performance Metric to use for choosing the optimisation result ("LV","RSS","RCM","Entropy")}
}
\value{
MSARM.fit allows its user to estimate Markov-Switching AR-Models.
The potential performance Metrics that could be used are:
"LV" -> Utilizes the estimated conditonal log-likelhood function for optima selection
"RSS" -> Utilizes the quality of the in-sample fit for optima selection
"RCM" -> Utilizes an Gini-Coefficient approach for optima selection
"Entropy" -> Utilizes an Entropy approach for optima selection
MSARM utilizes the EM-Algorithm outlined in Hamilton(1994) and Hamilton(1990).
Furthermore more explanation regarding the theory implemented and how it builds beyond Hamiltons work can be found in From Theory to Application: Developing MSARM, an R Package for Markov-Switching Autoregressive Models
}
\description{
MSARM.fit
}
\examples{
MSARM.fit(Y_T,K,N,Switcher,m = 1,threshold_value = 0.5, max_value = 250, R_value = 5, Crit_value = "LV")
}
