% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSARM_plot.R
\name{MSARM.plot}
\alias{MSARM.plot}
\title{MSARM.plot()}
\usage{
MSARM.plot(erg_MSARM.predict, conf = TRUE)
}
\arguments{
\item{erg_MSARM.predict}{output from MSARM.predict}

\item{conf}{if set to TRUE and MSARM.predict utilized boot = TRUE, then confidence intervals for the forecasts will be additionally plotted}
}
\value{
MSARM.plot allows its user to plot the forecasts from MSARM.predict to see how the time series is expected to behave in future. Furthermore
MSARM.plot allows its user if conf = TRUE and boot = TRUE (in MSARM.predict) to additionally plot the bootstrap confidence intervals for the forecasts
}
\description{
MSARM.plot()
}
\examples{
MSARM.plot(erg_MSARM.predict,conf = TRUE)
}
