% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSARM_predict.R
\name{MSARM.predict}
\alias{MSARM.predict}
\title{MSARM.predict
MSARM.predict allows its user to make forecasts utilizing an earlier estimated Markov-Switching AR model, the underlying theory is described in Müller (2025, page 24)}
\usage{
MSARM.predict(
  res_MSARM.fit,
  n.ahead = 1,
  boot = FALSE,
  levels = c(0.95, 0.9, 0.8, 0.7, 0.6),
  L = 10000
)
}
\arguments{
\item{res_MSARM.fit}{An object containing the estimation results of MSARM.fit}

\item{n.ahead}{Number of time periods to forecast ahead}

\item{boot}{Whether bootstrap forecasts and bootstrap confidence intervals should be implemented instead of standard forecasts}

\item{levels}{The levels of the boostrap confidence intervals, if boot is set to TRUE}

\item{L}{Number of boostrap estimations computed}
}
\description{
MSARM.predict
MSARM.predict allows its user to make forecasts utilizing an earlier estimated Markov-Switching AR model, the underlying theory is described in Müller (2025, page 24)
}
\examples{
MSARM.predict(res_MSARM.fit, n.ahead = 4) would give for a quarterly time series the forecast for the next year.
}
