% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_0_function.R
\name{P_0_function}
\alias{P_0_function}
\title{Title}
\usage{
P_0_function(Y_T, N, K)
}
\arguments{
\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
A random starting value of P is generated
}
\description{
Title
}
\examples{
P_0_function(Y_T,N,K)
}
