% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P_ij_next_function.R
\name{P_ij_next_function}
\alias{P_ij_next_function}
\title{Title}
\usage{
P_ij_next_function(alpha, P, Y_T, N, K)
}
\arguments{
\item{alpha}{alpha vector (intercept, coefficients and error variance) from the last iteration}

\item{P}{Transpose of the transition matrix Pi}

\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
Computes the P matrix for the next iteration, based on the last iteration
}
\description{
Title
}
\examples{
P_ij_next_function(alpha,P,Y_T,N,K)
}
