% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_1_0_function.R
\name{zeta_1_0_function}
\alias{zeta_1_0_function}
\title{Title}
\usage{
zeta_1_0_function(Y_T, N, K)
}
\arguments{
\item{Y_T}{Data}

\item{N}{Number of underlying Regimes}

\item{K}{Lag order of the AR-Model}
}
\value{
A random starting value of zeta (regime probability) is generated
}
\description{
Title
}
\examples{
zeta_1_0_function(Y_T,N,K)
}
