% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_1_0_function.R
\name{zeta_1_0_function}
\alias{zeta_1_0_function}
\title{zeta_1_0_function}
\usage{
zeta_1_0_function(Y_T, N, K)
}
\arguments{
\item{Y_T}{Data}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}
}
\value{
Creates a vector representing the random starting point for the optimal inference over the regimes
}
\description{
zeta_1_0_function creates zeta_1_0 as part of the random starting point for the EM algorithm
}
\examples{
zeta_1_0_function(Y_T,N,K)
}
