% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeta_Yt_function.R
\name{zeta_Yt_function}
\alias{zeta_Yt_function}
\title{zeta_Yt_function
zeta_Yt_function allows its user to get optimal inference over the regimes as described in Hamilton (1994, page 692)}
\usage{
zeta_Yt_function(alpha, P, Y_T, N, K)
}
\arguments{
\item{alpha}{alpha matrix from the last iteration}

\item{P}{P matrix from the last iteration}

\item{Y_T}{Data}

\item{N}{Number of underlying regimes}

\item{K}{Lag-order of the AR model}
}
\value{
Returns the optimal inference over the regimes
}
\description{
zeta_Yt_function
zeta_Yt_function allows its user to get optimal inference over the regimes as described in Hamilton (1994, page 692)
}
\examples{
zeta_Yt_function(alpha,P,Y_T,N,K)
}
